implement pair valuation

+++ contracts/interfaces/IBondingCalculator.sol
@@ -2,5 +2,8 @@
 pragma solidity 0.8.10;
 
 interface IBondingCalculator {
-  function valuation( address pair_, uint amount_ ) external view returns ( uint _value );
+  function valuation(address pair_, uint256 amount_)
+    external
+    view
+    returns (uint256 _value);
 }
